Pin CPU-only torch/torchvision wheels to skip the CUDA toolkit download
PyPI's default Linux torch wheel bundles the full CUDA toolkit (cublas,
cudnn, nccl, triton, etc. — 2GB+) as required dependencies even when no
GPU is present, which is always true for this Docker image: it has no
GPU, and YOLO_DEVICE=auto already falls back to CPU inference on its own.
That download bought nothing but build time.

Pinned torch==2.5.1+cpu / torchvision==0.20.1+cpu from PyTorch's own CPU
wheel index instead — same CPU-inference behavior, ~175MB total instead
of 2GB+. requirements.txt documents how to revert this for anyone who
does want GPU acceleration.

Also fixes 2 mypy errors this surfaced: motion_detection.py's
select_best_algorithm passed cv2.cvtColor/absdiff output straight to
np.mean/np.median, whose signatures in the *exact pinned* numpy==2.1.3
(as opposed to CI's newer, unpinned numpy) don't accept cv2-stubs'
MatLike union dtype. Not caught by CI, which never installs torch and
resolves a newer numpy than requirements.txt pins. Fixed with an
explicit .astype(np.uint8) (the arrays are already uint8 at runtime).

Verified: a clean `pip install -r requirements.txt` in a Python 3.12 venv
(matching the Dockerfile's python:3.12-slim base) installs in ~1m37s with
zero nvidia-*/cuda-toolkit/triton packages; ruff clean, mypy clean, and
the full 203-test suite passes with real (CPU) object detection enabled
(no ENABLE_OBJECT_DETECTION=false override needed).

+++ backend/app/services/motion_detection.py
@@ -317,10 +317,10 @@ def select_best_algorithm(
                 break
             if frame_index % sample_stride == 0:
                 gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
-                brightness_samples.append(float(np.mean(gray)))
+                brightness_samples.append(float(np.mean(gray.astype(np.uint8))))
                 if prev_gray is not None:
                     delta = cv2.absdiff(prev_gray, gray)
-                    noise_samples.append(float(np.median(delta)))
+                    noise_samples.append(float(np.median(delta.astype(np.uint8))))
                 prev_gray = gray
                 collected += 1
             frame_index += 1
